pyproject.toml
[project]
name = "orchestron"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.109.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.43",
    "youtube-transcript-api==0.6.2",
]
